+++ commands/status.js
@@ -1,0 +1,24 @@
+module.exports = {
+	category: 'Configuration',
+	description: 'Sets the bots status.',
+
+	minArgs: 1,
+	expectedArgs: '<status>',
+
+	slash: 'both',
+	testOnly: true, // This only works for test servers!
+	ownerOnly: true,
+
+	callback: async ({ client, text }) => {
+		client.user.setPresence ({
+			status: 'dnd',
+			activities: [
+				{
+					name: text,
+				}
+			]
+		});
+
+		return '✨ Status set to `' + text + '`!';
+	}
+}
