+++ data/retoEmojis.js
@@ -1,0 +1,4 @@
+module.exports = {
+    karmaEmoji: '<:karma:862440157525180488>',
+    confirmationEmoji: '<:retoolconfirm:1059560587417354270>'
+}
