--- data/retoEmojis.ts
@@ -1,4 +1,0 @@
-export const retoEmojis = {
-    karmaEmoji: '<:karma:862440157525180488>',
-    confirmationEmoji: '<:retoolconfirm:1059560587417354270>'
-}
