+++ index.js
@@ -1,0 +1,34 @@
+const { Client, Intents, Partials } = require("discord.js");
+const WOK = require("wokcommands");
+const path = require("path");
+require("dotenv/config");
+
+const client = new Client({
+	intents: [
+		Intents.FLAGS.GUILDS,
+		Intents.FLAGS.GUILD_MESSAGES,
+		Intents.FLAGS.GUILD_MESSAGE_REACTIONS
+	],
+	partials: [
+		'MESSAGE',
+		'CHANNEL',
+		'REACTION'
+	]
+});
+
+client.on("ready", () => {
+  new WOK(client, {
+		commandsDir: path.join(__dirname, 'commands'),
+		featuresDir: path.join(__dirname, 'features'),
+		messagesPath: path.join(__dirname, 'i18n/messages.json'),
+		testServers: [
+			'952707420700934195', // Retool Development
+		],
+		botOwners: [
+			'690962744454938734', // Honiemun
+		],
+		mongoUri: process.env.MONGO_URI
+	});
+});
+
+client.login(process.env.TOKEN);
