+++ package.json
@@ -1,6 +1,5 @@
 {
   "dependencies": {
-    "discord.js": "^13.10.2",
     "dotenv": "^16.0.1",
     "mongoose": "^6.3.0",
     "ts-node": "^10.9.1",
@@ -12,7 +11,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon index.ts"
+    "dev": "nodemon index.js"
   },
   "keywords": [],
   "author": "",
